rework package structure
