+++ .github/workflows/test.yaml
@@ -27,7 +27,7 @@ jobs:
         run: task build
 
       - name: Unit tests
-        run: task test-with-race
+        run: task test
 
       - name: Upload coverage reports to Codecov
         uses: codecov/codecov-action@v5
