+++ codecov.yml
@@ -1,0 +1,4 @@
+---
+coverage:
+  ignore:
+    - "main.go"
