+++ generator.go
@@ -1,24 +1,20 @@
-package generator
+package main
 
 import (
 	"fmt"
 	"os"
-
-	"github.com/weastur/replacer/internal/config"
 )
 
-func Run(cfg *config.Config) error {
+func Run(cfg *Config) error {
 	filePath := os.Getenv("GOFILE")
 	if filePath == "" {
-		fmt.Println("GOFILE environment variable is not set")
+		PrintVerbose("GOFILE environment variable is not set")
 
 		return nil
 	}
 
 	fileInfo, err := os.Stat(filePath)
 	if err != nil {
-		fmt.Printf("Failed to get file info: %v\n", err)
-
 		return fmt.Errorf("failed to get file info: %w", err)
 	}
 
@@ -26,8 +22,6 @@ func Run(cfg *config.Config) error {
 
 	content, err := os.ReadFile(filePath)
 	if err != nil {
-		fmt.Printf("Failed to read file: %v\n", err)
-
 		return fmt.Errorf("failed to read file: %w", err)
 	}
 
@@ -38,8 +32,6 @@ func Run(cfg *config.Config) error {
 
 	err = os.WriteFile(filePath, []byte(updatedContent), filePerms)
 	if err != nil {
-		fmt.Printf("Failed to write file: %v\n", err)
-
 		return fmt.Errorf("failed to write file: %w", err)
 	}
